Fix logging and test failures.
The server stress tests fail due to a lot of connection being in TCP_WAIT
state.  This could be fixed by reducing the TCP_WAIT timeout or by reusing
connections in TCP_WAIT state.  This alone is not enough to fix it.  If the rate
of connections is too high and the parallelism allows for more connections than
possible, it still fails.  To handle this, semaphores are introduced in the test
cases to control the number of connections.